+++ aws/autoscaling/common/variables.tf
@@ -1,4 +1,0 @@
-variable "entity" {}
-variable "min_capacity" {}
-variable "optype" {}
-variable "type" {}