+++ aws/dynamodb.tf
@@ -1,16 +1,10 @@
 resource "aws_dynamodb_table" "geopoiesis-scopes" {
-  name           = "${var.scopes_table_name}"
-  read_capacity  = "${var.min_capacity}"
-  write_capacity = "${var.min_capacity}"
-  hash_key       = "scope"
+  name     = "${var.scopes_table_name}"
+  hash_key = "scope"
 
   attribute {
     name = "scope"
     type = "S"
-  }
-
-  lifecycle {
-    ignore_changes = ["read_capacity", "write_capacity"]
   }
 
   tags {
@@ -19,19 +13,11 @@ resource "aws_dynamodb_table" "geopoiesis-scopes" {
   }
 }
 
-module "geopoiesis-scopes-autoscaling" {
-  source = "./autoscaling"
-
-  entity       = "${aws_dynamodb_table.geopoiesis-scopes.name}"
-  min_capacity = "${var.min_capacity}"
-}
-
 resource "aws_dynamodb_table" "geopoiesis-runs" {
-  name           = "${var.runs_table_name}"
-  read_capacity  = "${var.min_capacity}"
-  write_capacity = "${var.min_capacity}"
-  hash_key       = "scope"
-  range_key      = "id"
+  name         = "${var.runs_table_name}"
+  billing_mode = "PAY_PER_REQUEST"
+  hash_key     = "scope"
+  range_key    = "id"
 
   attribute {
     name = "scope"
@@ -68,8 +54,6 @@ resource "aws_dynamodb_table" "geopoiesis-runs" {
     range_key       = "id"
     name            = "byqueued"
     projection_type = "ALL"
-    read_capacity   = "${var.min_capacity}"
-    write_capacity  = "${var.min_capacity}"
   }
 
   global_secondary_index {
@@ -77,8 +61,6 @@ resource "aws_dynamodb_table" "geopoiesis-runs" {
     range_key       = "id"
     name            = "bytype"
     projection_type = "ALL"
-    read_capacity   = "${var.min_capacity}"
-    write_capacity  = "${var.min_capacity}"
   }
 
   global_secondary_index {
@@ -86,8 +68,6 @@ resource "aws_dynamodb_table" "geopoiesis-runs" {
     range_key       = "id"
     name            = "byvisibility"
     projection_type = "ALL"
-    read_capacity   = "${var.min_capacity}"
-    write_capacity  = "${var.min_capacity}"
   }
 
   global_secondary_index {
@@ -95,16 +75,6 @@ resource "aws_dynamodb_table" "geopoiesis-runs" {
     range_key       = "id"
     name            = "byworkerassigned"
     projection_type = "ALL"
-    read_capacity   = "${var.min_capacity}"
-    write_capacity  = "${var.min_capacity}"
-  }
-
-  lifecycle {
-    ignore_changes = [
-      "global_secondary_index",
-      "read_capacity",
-      "write_capacity",
-    ]
   }
 
   tags {
@@ -113,69 +83,18 @@ resource "aws_dynamodb_table" "geopoiesis-runs" {
   }
 }
 
-module "geopoiesis-runs-autoscaling" {
-  source = "./autoscaling"
-
-  entity       = "${aws_dynamodb_table.geopoiesis-runs.name}"
-  min_capacity = "${var.min_capacity}"
-}
-
-module "geopoiesis-runs-autoscaling-byqueued" {
-  source = "./autoscaling"
-
-  entity       = "${aws_dynamodb_table.geopoiesis-runs.name}/index/byqueued"
-  min_capacity = "${var.min_capacity}"
-  type         = "index"
-}
-
-module "geopoiesis-runs-autoscaling-bytype" {
-  source = "./autoscaling"
-
-  entity       = "${aws_dynamodb_table.geopoiesis-runs.name}/index/bytype"
-  min_capacity = "${var.min_capacity}"
-  type         = "index"
-}
-
-module "geopoiesis-runs-autoscaling-byvisibility" {
-  source = "./autoscaling"
-
-  entity       = "${aws_dynamodb_table.geopoiesis-runs.name}/index/byvisibility"
-  min_capacity = "${var.min_capacity}"
-  type         = "index"
-}
-
-module "geopoiesis-runs-autoscaling-byworkerassigned" {
-  source = "./autoscaling"
-
-  entity       = "${aws_dynamodb_table.geopoiesis-runs.name}/index/byworkerassigned"
-  min_capacity = "${var.min_capacity}"
-  type         = "index"
-}
-
 resource "aws_dynamodb_table" "geopoiesis-workers" {
-  name           = "${var.workers_table_name}"
-  read_capacity  = "${var.min_capacity}"
-  write_capacity = "${var.min_capacity}"
-  hash_key       = "worker_id"
+  name         = "${var.workers_table_name}"
+  billing_mode = "PAY_PER_REQUEST"
+  hash_key     = "worker_id"
 
   attribute {
     name = "worker_id"
     type = "S"
   }
 
-  lifecycle {
-    ignore_changes = ["read_capacity", "write_capacity"]
-  }
-
   tags {
     Name        = "Owner"
     Description = "Geopoiesis"
   }
-}
-
-module "geopoiesis-workers-autoscaling" {
-  source = "./autoscaling"
-
-  entity       = "${aws_dynamodb_table.geopoiesis-workers.name}"
-  min_capacity = "${var.min_capacity}"
 }